Correct error checking logic in DiscardOtherRebootReason & DiscardCommonRebootReason functions.
